implement base classes for credentials and request sessions
